character_generation/requirements.txt
google-cloud-storage==2.14.0
google-cloud-aiplatform==1.38.0
